Update rustyline to latest
