+++ crates/nu-cli/src/shell/completer.rs
@@ -3,7 +3,7 @@ use crate::completion::flag::FlagCompleter;
 use crate::completion::path::{PathCompleter, PathSuggestion};
 use crate::completion::{self, Completer, Suggestion};
 use crate::context;
-
+use std::borrow::Cow;
 pub(crate) struct NuCompleter {}
 
 impl NuCompleter {}
@@ -109,8 +109,8 @@ fn select_directory_suggestions(completed_paths: Vec<PathSuggestion>) -> Vec<Pat
         .collect()
 }
 
-fn requote(value: String) -> String {
-    let value = rustyline::completion::unescape(&value, Some('\\'));
+fn requote(orig_value: String) -> String {
+    let value: Cow<str> = rustyline::completion::unescape(&orig_value, Some('\\'));
 
     let mut quotes = vec!['"', '\'', '`'];
     let mut should_quote = false;
